Add Graphite example configuration
Add graphite example configuration based on the original bosh-lite
config:

https://github.com/CloudCredo/graphite-statsd-boshrelease/blob/75e1a6794712aa2f84ac3c3280b977b7ed7fbb33/examples/graphite-bosh-lite.yml

We define some defaults for the AWS configuration, using large_z1 type
for the VM with SSD disks to keep with the load of the number of
metrics of a CF deployment like ours.

In order to specify cloud_properties for a persistent disk it's
necessary to use disk pools[1], so this PR adds a disk pool for the
graphite persistent disks, and updates the job to use it.

[1] https://bosh.io/docs/persistent-disks.html