Initial commit to create bastion host
This creates bastion host and configures hostname correctly. Also we
now use user ssh keys which are uploaded to AWS by terraform. NAT was
removed.

We have decided to put dev as a domain prefix for a time being.
We can change it later, add more domains, parametrize it etc.