Add a trivial build script and hook up to specs.
The script currently does nothing but cat the file. This is to allow the
specs to be written to consume the output from a script. In future the
script will be extended to use spruce to merge the templates together,
and the tests won't need to change how they load the manifest.