Add tf-destroy.sh script
This script is used by tasks in destroy-cf pipeline. We split this out
from the main pipeline and task definition, because the script can be
used by multiple different tasks and on it's own. This is also to
demonstrate a different and usual pipeline organization. With the
destroy pipeline, we include tasks from yml files, which run scripts
already present in this repository, removing the need to inline the
script in the pipeline or task definition.