Add tf-destroy task
This task is used by multiple jobs in the destroy pipeline. The task
definition by itself is not complete and this is on purpose. The missing
field is params (required to set environment variables for tf-destroy
script). This is because there is no good default set of parameters and
fly only expands parameters from variables in the pipeline and not task.
Hence `params` is defined in the pipeline.

Another catch with including a task from file in pipeline is that you
can't simply add members/items into yml hash or list. E.g. when I want
to use `paas-cf` git repo and some S3 bucket as inputs, I can't simply
define `paas-cf` in the task in inputs and the S3 bucket resource in
the pipeline. The pipeline hash definition overrides whole hash/list in
task. There is no merging. Hence I need to include `paas-cf` again
explicitly in the pipe definition.

This is to highlight the default/recommended organization of items in
concourse. We might want to use spruce to merge these files into one
big pipeline definition...