Deploy script: don't fly watch automatically
Fly outputs state of the last successful job or the job in progress.
When you deploy it is quite likely that you run a deploy job before,
hence you get output of the last job instead of the build you're
starting just now. Add a note that user should likely wait a bit
before executing the fly command themselves.