Copy original bosh templates and add example stubs
Copy the original bosh templates and add example stubs (secrets and
terraform-outputs) to spruce them up.

Also generate a reference bosh manifest based on the original manifests
in `./manifests/bosh/*`, which will allow test the refactoring.

The file was generated with:

    pruce merge \
    --prune meta \
    --prune secrets \
    --prune terraform_outputs
	<(
	spiff merge \
	    manifests/templates/bosh/bosh-template.yml \
	    manifests/templates/aws/bosh/bosh-manifest.yml \
	    bosh-manifest/deployments/example-stubs/bosh-secrets.yml \
	    bosh-manifest/deployments/example-stubs/terraform-outputs-aws.yml
	) > bosh-manifest/reference-bosh-manifest.yml

We add a script to generate bosh template with spruce, which  will also prune
the `meta`, `secrets` and `terraform_outputs` root keys.