Split cloud_provider and pkg-cache in bosh manifest
Move cloud_provider config to separate file

This configuration is specific for bosh-init. As described in bosh
documentation, https://bosh.io/docs/deployment-manifest.html:

   cloud_provider [Hash, required]: Specifies CPI configuration for the bosh-init to create VMs, etc. Regular deployment manifests cannot specify this block.

We also move the shared properties in the AWS files to this file.

Secondly, move bosh compiled pkg cache config in separated file. This way it
would be easier to disable that feature.