Refactor deploy-pipeline to get a generic varsfile
Avoid hardcoding variables in the deploy-pipeline.sh so that we can
resuse for other pipelines which require different variables.

Instead, we pass a yaml file to be loaded with `--load-from-vars`
option.

In the `create-deployer.sh` script, we generate the vars file, and pass
it as a file buffer to avoid generate any temporary file.

We can reuse this pattern later.