Upgrade BOSH and AWS CPI for IAM profiles support
In the case of BOSH, the v208 supports IAM profiles configuration.
But we use the latest version v236 because we are starting with
a complete new setup.

In the case of AWS CPI, we upgrade to the latest version, which also
supports the IAM profiles and provides several bugfixes.

The new AWS CPI release renames the cpi job to `aws_cpi`, so we need to
change some configuration in the bosh deployment to keep everything
working. For instance: `director.cpi_job` which renders the path to the
CPI driver.