Setup AWS IAM profiles for BOSH deployment
Implement IAM profiles for AWS authentication instead of use of
plain AWS credentials, as described in https://bosh.io/docs/aws-iam-instance-profiles.html

We remove the references to unnecesary AWS credentials and we setup bosh
to use the IAM profile `bosh-director`. This IAM profile must be
setup to be able to create ec2 resources and pass a IAM roles if required
(for example, CF API must be able to read S3 assets)

As this pipeline can be executed from local desktops, we leave open the
option of specifying AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY
environment variables.

These environment variables populate the
pipeline variables `aws_access_key_id` and `aws_secret_access_key`,
which will be also used to set the environment variables in the running
tasks.

This way, the pipeline can be run populating the AWS credentials, if it
is run from a local desktop, or without populating them, if it is run
from a AWS server with a IAM profile (e.g. bosh-bootstrap). In each case:

 * If the AWS credentials are populated:
   * Terraform: it will use them directly.
   * Bosh-init: we override the AWS configuration in the manifest
     related to `cloud_provider` only, but not to the deployed bosh job.
     This way `bosh-init` will use AWS creds but bosh will use his IAM
     profile.

 * If the AWS credentials are not populated.
   * Terraform: it will try to get the credentials from the IAM profile.
   * Bosh-init: We will not override the default manifest, which is
     setup to use IAM profiles for bosh-init and for the bosh
     deployment.

Example policies assigned to the roles are as follows:

 * `bosh-bootstrap`: for the VM that can run this very pipeline:

    {
    "Version": "2012-10-17",
    "Statement": [
    {
        "Action": "ec2:*",
        "Effect": "Allow",
        "Resource": "*"
    },
    {
        "Action": "route53:*",
        "Effect": "Allow",
        "Resource": "arn:aws:route53:::hostedzone/Z3SI0PSH6KKVH4"
    },
    {
        "Effect": "Allow",
        "Action": "elasticloadbalancing:*",
        "Resource": "*"
    },{
        "Effect":"Allow",
        "Action":"iam:PassRole",
        "Resource": [
            "arn:aws:iam::<accound_id>:role/bosh-director"
        ]
    },{
        "Effect": "Allow",
        "Action": [ "s3:*" ],
        "Resource": [
        "arn:aws:s3:::*"
        ]
    }]
    }

 * `bosh-director`: for the bosh director:

    {
    "Version": "2012-10-17",
    "Statement": [{
        "Action": "ec2:*",
        "Effect": "Allow",
        "Resource": "*"
    },{
        "Effect": "Allow",
        "Action": "elasticloadbalancing:*",
        "Resource": "*"
    },{
        "Effect":"Allow",
        "Action":"iam:PassRole",
        "Resource":"arn:aws:iam::<accound_id>:role/cf-cloudcontroller“
    },{
        "Effect": "Allow",
        "Action": [ "s3:*" ],
        "Resource": [
        "arn:aws:s3:::*-cf-droplets",
        "arn:aws:s3:::*-cf-droplets/*",
        "arn:aws:s3:::*-cf-buildpacks",
        "arn:aws:s3:::*-cf-buildpacks/*",
        "arn:aws:s3:::*-cf-resources",
        "arn:aws:s3:::*-cf-resources/*",
        "arn:aws:s3:::*-cf-packages",
        "arn:aws:s3:::*-cf-packages/*",
        "arn:aws:s3:::shared-cf-bosh-blobstore*"
        ]
    }]
    }