Only allow concourse to connect to bosh
Allow only concourse to connect to BOSH. We need to get the security
group of the deployed concourse from the terraform output, and
import it in the terraform configuration for bosh.

We do not allow the office to connect anymore.

Instead of connecting directly to the public IP, we setup bosh-init to
connect only to the private ip.