Pin concourse/lite version to 0.69.0
To ensure that we are all using the same version of Concourse. This is the
latest version box available from Atlas and I have tested it with the
`create-deployer` pipeline in `master`.

- https://atlas.hashicorp.com/concourse/boxes/lite

On important change, if you were using 0.68.0 previously, is that `fly
hijack` now sets the correct PWD and environment variables for a job. Which
makes debugging failed jobs much easier.