Randomly generate bosh secrets in a job
Add a job to generate a file with a group of random passwords required
for bosh deployment, instead of using the example file in the
repository.

This job will try to download the secrets file from a bucket,
creating a new one it if it is missing only the first time.

Note: there is a story to refactor and implement this consistently
across the deployment.