Implement a init job to trigger the bosh pipeline
Implement a entry "init" job as entry point of the pipeline.

This job allows us clearly define the entry points and triggers of the
pipeline (in this case, paas-cf) and provides a clear job name to
work as a interface to start the pipeline using the concourse pipeline.

We also add a fake resource `pipeline-trigger` to trigger the next jobs.