Travis tests bosh manifest generation
Update travis definition to also run rspec on the bosh manifest.

Travis only supports one language per repository, but the terraform
tests can run with the ruby language, which includes bash.

We change matrix to includes bosh tests for terraform and rspec, and
we change the script to conditionally run a terraform show or
rspec with bundle.