Add vagrant concourse lite deploy script
Deploys AWS Vagrant bootstrap concourse lite, sets up auth and adds
create-deployer and destroy-deployer pipelines.

Some automatic passwords are generated for concourse DB and concourse interface.
For that we use a hash of the current AWS credentials, the environment name and
the service id.

This allows create secure, unique and repeteable passwords.