Remove use of terraform remote in create-deployer
Instead use concourse resources to read and write the state file. This
will avoid the race condition between terraform writing to the state in
one job, and the next job needing to read that updated state.

The only remaining place where these state files are not managed by
concourse is in the init-bucket job. This uses the s3init.sh script to
create blank state files if they don't already exist. This step can't
use a concourse resource because it needs to conditionally write. This
should not be subject to race conditions because S3 privodes
read-after-write consistency for PUTS of new objects [1].

[1]https://aws.amazon.com/s3/faqs/#What_data_consistency_model_does_Amazon_S3_employ