Inline tf-destroy-vpc task config.
This makes it consistent with how all the other tasks are defined.

This also removes and inlines the tf-destroy.sh script in favour of
inlining the scripts. Again, this is done for consistency with other
tasks. Additionally, this makes it more obvious where the updated stste
file is written to so that it can be written back to S3 by concourse in
upcoming commits.