Use a resource to write terraform state for state bucket
This was the last thing using terraform remote.

Note: This task will fail if run more than once because it doesn't read
the state file before running. It's not possible for it to read the
state file beforehand because the bucket that holds the state file
doesn't necessarily exist. This behaviour was the same with the previous
implementation using terraform remote.