Add vagrant_cidr into ssh and concourse SGs
The new AWS bootstrap-concourse Vagrant box needs to be able to connect
to the deployer concourse in eu-west-1. For this, add vagrant_cidrs
variable. Empty by default and configured at runtime when AWS vagrant
IP is known. To handle empty IP, we used concat and compact TF
functions, to create a valid list of SG cidrs. We chose this approach
instead of using some dummy default IP (e.g. 127.0.0.1).