Remove vagrant IP from SGs after deployment
Remove Vagrant IP from ssh and concourse security groups after concourse
is deployed. Vagrant no longer needs to access it, unless it runs the
deployment again. In that case it will temporarily enable it's access
again. This is so that we don't leave the vagrant IP in the SG even
after we destroy (or stop) the Vagrant box and lose the public IP,
which then gets used by someone else.