Fetch bucket state file if it exists.
If terraform runs without a state file, but the bucket already exists,
it will error with a conflict, and write out a state file with no
resources in it because it hasn't created anything itself. We don't want
to write this empty state file to S3 because that will overwrite the one
that contains details of the bucket.

Instead, we add a task to attempt to fetch the state file if it's
present. This means that terraform will use this file when the bucket
exists, but on the initial run, it won't error or stall when there is no
bucket to hold the state file yet.