Stop vagrant using the insecure deployer key
The AMI we are using contains the default vagrant key -
https://github.com/mitchellh/vagrant/blob/master/keys/vagrant.pub in its
`/home/ubuntu/.ssh/authorized_keys` file. Vagrant will detect this
insecure public key and generate a new public key which is stored on the
host machine in
`vagrant/.vagrant/machines/default/aws/private_key`. This means that
access with `vagrant ssh` will work correctly.

There is a small window of opportunity where we are running an instance
with the insecure key - however we are protected by a security group
limiting access to office ips.

We still need to override the ssh username as the ami does not contain a
vagrant user account.

This only stops vagrant using the insecure deployer key - we still need
to do work to stop the deployer concourse and microbosh using the
insecure deployer key