Upgrade deployer Concourse to 0.70.0
In order to pull in the fix for concourse/concourse#167 which made
`bosh-init` hang when unmounting disks from the "Deployer Concourse"
instance if there were any changes to the manifest.

This now matches the version of concourse/lite that we're using for the
"Bootstrap Concourse". We shouldn't be affected by any of the
backwards-incompatible entries in the changelog, but you will want to run
`fly sync`:

- http://concourse.ci/release-notes.html#v0.70.0