Remove insecure deployer for bosh-init
Bosh init needs ssh to create a tunnel for the agent on the stemcell
bosh init create to connect to the director running on bosh init. To do
this it requires a keypair. However, it only requires this keypair
during the initial deploy, as a redeploy will recreate the instance. So
it is safe to generate a new keypair each time we run a deployment. It
is also safe to run terraform without saving the state file between runs
- due to the dummy state file, terraform will see the resource as
requiring changes and destroy and recreate it.

For the case that we need to ssh to the bosh init director without
redeploying, we upload the key to s3 so it can be retrieved if needed.
The role given to the concourse vm running bosh init should only allow
write access to the bucket.