Generate SSH key and upload to AWS in VPC creation
Add a task to generate a SSH key for all the environment and upload it
using terraform.

The task will generate a key, but upload to S3 only if there it was not
uploaded before. This way we ensure that the key is generated once.

In terraform, we create the SSH key pair using the public key and output
its name to be used later with terraform.

FIXME: we need to install openssh to the curl-ssl container or create a new one.