Fix concourse ephemeral disk definition
For some reason, our manifest generation was losing ephemeral disk
definition when it was defined after ELB. When we move the disk
definition to where it is now, template is rendered correctly and
ephemeral disk info is included in the manifest. This way concourse
really gets 40G ephemeral disk. 40G should be well enough, as
ephemeral disk doesn't seem to be used too much.