A set of NAT gateway resources
This commit creates all required resources to create NAT gateway.
NAT GW is placed in infra subnet that has internet access. It has its
own EIP and modifies CF subnet routing tables to accept outgoing
traffic.

Internet routing table that was created by VPC job was not used on that
stage. I have moved it to CF Terraform run as it is needed here to be
associeded with subnets with NAT gateways