Deploy pipeline downloads all releases for diego
We download all the releases from Diego. In this case we replicate the
approach of download using curl, but using multiple tasks to download in
parallel.

The bosh-deployment release will upload them.

Note that the resource will crash if the manifest does not include the
uploaded releases as per
https://github.com/concourse/bosh-deployment-resource/issues/18

Also note that this resource does change the manifest to force the
version of the releases to be the same than the ones specified in the
put action.