Split post_deploy script in multiple scripts
For easier maintenance we split the `concourse_post_deploy.sh` in
separated scripts per topic. All the scripts are located in
`vagrant/post-deploy.d` and loaded programatically from the Vagrantfile as
independent shell providers.

Any arguments or configuration will be passed as environment variables
instead of arguments. In this case, we only need to pass the environment
variables for concourse basic authentication.