Remove aws-secrets.yml from cloudfoundry manifest.
There's no longer anything in the manifest that uses these secrets. This
therefore removes them from the manifest building script and tests. It also
removes the pipeline step to generate them etc.

This removes the last thing in the pipeline that was using access keys,
so this also removes the code passing these values to the pipeline.