Add a job to delete cf deployment at night
We want to automatically delete the CF deployment nightly to save money.

For that, we add a new task which is triggered by a time resource
scheduled to run every 24h at night.

We also add a time resource to mark a timestamp of the deployment, which
is updated regardless `deploy` is successful or not (using `ensure`).
This allows us to define a dependency between the `deploy` job and
the `delete` job.