deployer: Remove comments about starting/watching
The `curl` command to automatically start a job was commented out in
8c760e9. The other pipelines never had this and everyone has been starting
their jobs from the web UI.

The note that is printed afterwards, describing how to watch the job output
from the CLI, is no longer relevant now that we're not starting jobs
automatically. If you've started a job from the web UI then you'll see
output there.