Remove dependency between cf deploy and delete
Concourse dependency semantics only allow add a dependency between
jobs in the case the first job is successful (`passed: [ <jobs> ]`).

But we want the job `delete` to be run regardless of the final result
of the `deploy` job.

Because this, we decided to drop the dependency. The destroy job will be
triggered indepently and delete the environment overnight, although the
pipeline representation does not represent the logical dependencies.