Move unpause-pipeline command to common script
So that we have less duplication between these individual scripts. The
variables that are being removed are only used by the unpause command.

It also fixes a variation in the `deploy-cloudfoundry` script which calls
`fly` instead of `${FLY_CMD}` that causes the unpause to fail if you've
followed the instructions and only have `fly` installed in your current
working directory (by `vagrant/deploy.sh`):

    ➜  paas-cf git:(master) ✗ ./concourse/scripts/deploy-cloudfoundry.sh
    Concourse API target dcarley
    Concourse API http://192.168.100.4:8080
    AWS Region
    Deployment dcarley
    Pipeline deploy-cloudfoundry
    Config file /Users/dcarley/projects/paas/paas-cf/concourse/scripts/../pipelines/deploy-cloudfoundry.yml
    apply configuration? [yN]: y
    configuration updated
    ./concourse/scripts/deploy-cloudfoundry.sh: line 40: fly: command not found