Add pipeline to self-terminate concourse-lite
Add a pipeline for the vagrant concourse-lite which will be triggered
at night and terminate the VM.

It uses metadata to learn the VM region and ID and the awscli tools to
request it to terminate. As concourse-lite has a IAM role which allows
ec2 operations, the awscli do not require AWS credentials.