Use separated pipeline for nightly CF delete
To keep more clear the separation and make it easier to disable this
feature in some environments (e.g. production), we deploy a separated
pipeline that will autodelete the cloudfoundry deployment.

We add some logic to optionally disable this feature by setting
`DISABLE_AUTODELETE` environment variable.

The autodelete pipeline is uploaded in the same script than script that
uploads the deploy-cloudfoundry pipeline, so we can ensure that it is
always setup.