Include shellcheck in travis run.
In order to be able to install shellcheck on the container (without
requiring sudo and therefore not being able to run on the container
VMs), we have to add it from a repository that's in Travis' whitelist.
Shellcheck isn't available in the Ubunto repos for Precise, so we're
using the debian sid repos here instead.

We did consider installing from source, but that would require a haskell
compiler, and pulling down all the dependencies from the Internet, which
would take significantly longer.