Introduce Makefile and remove matrix build
Adding a Makefile makes it easier to run tests locally before pushing.

This also removes the matrix build from Travis because it wasn't gaining
us any performance. The slow part of the process is the initial install
step (ruby, packages etc), not the tests themselves. It was actually
hurting us running them in parallel - typical build time up to now is
3-5mins, but when running as a single job, it's in the 1m30s range.