Fix shellcheck errors for s3init.sh and s3get.sh
SC2046: Quote this to prevent word splitting.
SC2086: Double quote to prevent globbing and word splitting.

Disable check for SC2059: Don't use variables in the printf format
string. In this case we need to use $string as a format string because
we want printf to interpret the `\n` chars as newlines.