README: Standardise BOSH/CF section headings
To match the changes to the Bootstrap Concourse and Deployer Concourse
sections above that were changed in 5c736a3.

We can refer to the previous sections to avoid duplicating prerequisites.

I've deleted the description of the MicroBOSH pipeline and I'm going to add
this back in a subsequent commit to match the descriptions of Bootstrap
Concourse and Deployer Concourse.