Add a job to run acceptance-tests after CF deploy
Run the acceptance_tests errand using the bosh cli on the deployed
cloudfoundry. We use the manifest from the generate-manifest job, only
if the deployment was successful.

The job is not triggered automatically, as the acceptance tests are
heavier and slower than smoke tests.

We download the logs to allow the developer "hijack" the build
container to check the logs in case the tests fail.