README: Move and reword BRANCH env var
Move this to the top of the notes section so that it's more prominent. We
use this a *lot* when reviewing pull requests.

Reword it to make it clear that it applies to all of the pipeline scripts.

I've also removed the `export` statement and changed the example to passing
the environment variable to a script. This is intended to prevent people
from exporting it and forgetting about it.