Generate Concourse cert in pipeline
We want to stop using hard coded certificate and instead generate them and store
them in the environment S3 bucket. This way we don't need to maintain a central
secrets repository, it is much safer if each environment has its own secrets and
it creates a workflow that automates the renewal of these secrets.
Details:

* add a new S3 resource concourse-cert
* init with empty tar file if necessary
* generate certificate and make it available to Terraform
* upload certificate via S3 resource