Combine create/destroy MicroBOSH pipeline scripts
So that we only need to maintain and run a single script to configure the
pipelines for this component.

The "prerequisites" section seems like the most appropriate place to
document how to deploy them. However I think that we could later wrap all of
the pipelines for the Deployer Concourse in a single outer script, in much
the same way that `vagrant/deploy.sh` does for Bootstrap Concourse.

There are the following differences between the two scripts:

- pipeline name is now handled by a loop

Diff prior to this commit:

```diff
--- concourse/scripts/create-microbosh.sh	2016-02-01 10:50:21.000000000 +0000
+++ concourse/scripts/destroy-microbosh.sh	2016-02-01 10:50:21.000000000 +0000
@@ -4,8 +4,8 @@
 SCRIPT_DIR=$(cd "$(dirname "$0")" && pwd)

 env=${DEPLOY_ENV-$1}
-pipeline="create-microbosh"
-config="${SCRIPT_DIR}/../pipelines/create-microbosh.yml"
+pipeline="destroy-microbosh"
+config="${SCRIPT_DIR}/../pipelines/destroy-microbosh.yml"

 [[ -z "${env}" ]] && echo "Must provide environment name" && exit 100

```