README: Reword sharing Bootstrap Concourse section
Remove the description of the tunnel because it duplicates the description
under the Bootstrap Concourse / Deploy section above. I don't think we need
to describe the specifics of the tunnel, such as hostnames and ports,
because the URL remains the same from when you're setting up your own.

Change the section title because this is really only about sharing and not
SSHing for shell access.