CATs include routing tests but not routing api
This package contains routing specific acceptance tests, and GoRouter
(Context path, sticky sessions...).

We skip the routing api and routing services related tests, which require
additional services and are still in development. In later acceptance tests
these tests are in a separated suite called `router_api` but currently we
need to skip them using a regex.

Note that although the option `include_route_services` was previously
set to "true", these tests were not being run because they are part of the
suite `routing`, which was completely disabled. Because that,
`include_route_services=true` had not effect until unless we set
`include_routing=true`.