deploy-cf: Generate bbs_server certificates
Add logic to generate the bbs_server certificates using the Bosh CA
certificates.

The implementation will put a empty tar.gz file in the S3 bucket during
initialisation. The certificate will be generated only if that file is empty,
a new certificate will be generated and uploaded.

After that, a task will export the certificates in stub .yml files that
will be consumed by the generate-manifest job.