destroy-microbosh: BOSH delete trigger Terraform
The `bosh-terraform-destroy` job in the `destroy-microbosh` pipeline wasn't
being triggered after `bosh-init-delete` had completed. This left you with a
partial environment, until you manually trigger the `bosh-terraform-destroy`
job manually. This was noted in the "pipelines improvements" story and has
surprised me a few times.

We're going to have a story later to look at the `init` and trigger patterns
in all pipelines. This commit just concentrates on making it work, rather
than standardising any of the patterns currently in use.