deploy-cf: make generate-cf-cert more generic
As we are going to generate several certificates, we change the
generate-cf-certs implementation so that it will iterate over a list
of certificates to generate.

It checks if the .tar.gz is in the S3 state bucket using s3init without
arguments, and only generates and uploads the certificate if it is
missing.