s3init.sh: Do not upload if there is not template
The certificate generation is a expensive operation which requires entropy,
if there is not enough, it will hang. Because that we want to check
if the remote file is missing before generate the certificate.

In order to do that, we change the logic in s3init to only check and
get the file if no template is passed.