deploy-cf: Generate uaa_jwt_signing cert
Generate the uaa_jwt_signing key. In this case we only need the key, not
the certificate, but we generate it with the same logic.

In an additional task the public part of the key is extracted, so it can be
used as vertification key for the UAA tokens.

More info: https://github.com/cloudfoundry/uaa/blob/master/docs/Sysadmin-Guide.rst