Enable parallel job updates
* Remove per-job bosh update directives and use global settings.

* Set 2 parallel updates per job (using max_in_flight bosh directive)
We should update this later based on actual number of instances per
job we will have in production.

* Disable serial updates. This enables bosh to deploy jobs in parallel.

* We also disable canaries. This way it's only max_in_flight that
controls how many instances of a job are updated in parallel.

* Still deploy consul_z1 job first (both instances in parallel). This
is to enable other jobs to cleanly connect to consul immediately when
starting deployment. The jobs that need to do this would have to wait
for consul server to become available anyway, so deploying it as a 1st
job is not a slowdown.