deploy-cf: Refactor build-manifest.sh
With the current implementation build-manifest.sh for the cf manifest,
we need to modify the file everytime we want to add a new stub file.

We refactor the script to get all the stubs as arguments, so we can
append them directly. This simplifies the pipeline and the test.

With this change we don't need the dummy file for the certs.

Now the script is very similar to the one used to generate the
bosh-manifest.