Add custom-acceptance-tests job to pipeline.
Unlike the existing test jobs, these run in a container on the deployer
concourse, and not via a bosh errand. Given this code is defined in our
repository, and there is no bosh resource for it, there is no reason to
use an errand for there. Additionally, it's a lot quicker to run them if
we don't need to spin up a separate VM each time.

This includes a task to generate acceptance test config from the
cf-manifest.