README: Replace use of s3get with aws-cli
We can use the official aws-cli tool to do the same job as `s3get.sh`. If we
also replace the use of the script in the pipeline, then we can remove this
script from our repo.

I've included a very brief description (with some links) about how to
install and configure it.