create-deployer: Replace s3get with aws-cli
Use the official aws-cli utility instead of our custom script. We hope to
replace this task with s3-resource, if we can add support for it to error
instead of block when the bucket doesn't exist:

- https://www.pivotaltracker.com/story/show/110953182

In the meantime though, replacing this means that we have less code to
maintain. This is especially important as we don't currently have any tests
for this functionality.

We originally created the aws-cli container for 5448cee.