try-fetch-bucket-state: explicit error handling
Add the `-e` flag to the `try-fetch-bucket-state` task to exit early on
errors, as used in all other tasks. Then add `|| true` to the command which
attempts to fetch the state from the bucket and we expect to fail on the
very first run. This has the same effect but makes error handling more
explicit to read.