Use consul for etcd DNS resolution
Avoid using static ips for clients to connect to etcd. This only impacts the way the
clients connect to etcd, not the etcd cluster itself. Diego was already doing it, we
now apply the same to loggregator. This includes all metron agents.
All jobs must be updated to use consul agent as local DNS resolver.