Add the deployer-concourse EIP to the cf security group
In order to run the acceptance tests, the concourse machine needs to be
able to talk to the CF API.

If concourse is switched to using a NAT gateway for external
connections, this change can be reverted.

This required modifying the destroy-concourse pipeline so that the
concourse tfvars are used, otherwise it errors because the concourse IP
isn't defined.