Deterministically delay deletion for between 0-20m
This deterministically sleeps for a period of time between 0-20mins in order to
prevent all our deletion jobs running at the same time.

`short` exists because busybox `sh` cannot deal with enormous hex values when
converting to decimal. The decimal value generated may be negative, so
`${decimal##-}` pretends to be `abs` by stripping the `-`. The final sleep time
is anywhere between 0 and 20 minutes.